Delete Post is working!!!

+++ views/v_posts_delete.php
@@ -23,10 +23,10 @@
  <!-- 	print_r($posts);
 -->
 
-
-	<h2>Are you sure you want to delete this post?</h2>
-
-	<a href='/posts/p_delete/<?=$post_id?>'>
+	<p>Are you sure you want to delete this post? If not, click <a href="/posts">here </a>to return.</p>
+	
+	<a href='/posts/p_delete/<?=$post_id?>'><button type="button" class='button' id='confirmDelete'>Delete</button></a>
+	
 <!--
 		<form method='POST' action='/posts/p_delete'>
 			<input type="hidden" name='post_id' value='<?=$post['post_id']?>'>
